# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "رسالة الخيرية API",
    "description": "## نظام إدارة رسالة الخيرية\n\n### الأدوار المتاحة:\n| الدور | طريقة الدخول | الصلاحيات |\n|-------|-------------|----------|\n| **Donor** | Phone Number | تسجيل — عرض بياناته — الاشتراك |\n| **Reception** | Username | إدارة المتبرعين — تسجيل تبرعات |\n| **Admin** | Username | كل الصلاحيات + إدارة الموظفين |\n\n### كيفية الاستخدام:\n1. سجّل دخول من `/api/v1/auth/login`\n2. انسخ الـ `token` من الـ Response\n3. اضغط **Authorize** في الأعلى والصق: `Bearer {token}`",
    "contact": {
      "name": "Resala Dev Team",
      "email": "dev@resala.org"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/subscriptions": {
      "post": {
        "tags": [
          "Subscriptions — Donor"
        ],
        "summary": "[Donor] الخطوة 1 — إنشاء اشتراك في كفالة",
        "description": "الخطوة الأولى: إنشاء الاشتراك فقط.\r\n            \r\nبعده استخدم `POST /subscriptions/{id}/submit-payment` لتقديم الدفع.\r\n            \r\n**دورات الدفع (PaymentCycle):**\r\n- `1` = شهري\r\n- `3` = كل 3 شهور\r\n- `6` = كل 6 شهور",
        "operationId": "Subscriptions_Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/subscriptions/{id}/submit-payment": {
      "post": {
        "tags": [
          "Subscriptions — Donor"
        ],
        "summary": "[Donor] الخطوة 2 — تقديم طلب دفع",
        "description": "يتم بعد إنشاء الاشتراك.\r\n            \r\n**طرق الدفع:**\r\n- `1` VodafoneCash → `receiptImage` + `senderPhoneNumber`\r\n- `2` InstaPay → `receiptImage` + `senderPhoneNumber`\r\n- `3` Branch → `slotId` + `donorName` + `branchContactPhone`\r\n- `4` Representative → `deliveryAreaId` + `contactName` + `contactPhone` + `address`",
        "operationId": "Subscriptions_SubmitPayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "PaymentMethod": {
                    "$ref": "#/components/schemas/PaymentMethod"
                  },
                  "Amount": {
                    "type": "number",
                    "format": "double"
                  },
                  "ReceiptImage": {
                    "type": "string",
                    "format": "binary"
                  },
                  "SenderPhoneNumber": {
                    "type": "string"
                  },
                  "DeliveryAreaId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "ContactName": {
                    "type": "string"
                  },
                  "ContactPhone": {
                    "type": "string"
                  },
                  "Address": {
                    "type": "string"
                  },
                  "RepresentativeNotes": {
                    "type": "string"
                  },
                  "SlotId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "DonorName": {
                    "type": "string"
                  },
                  "BranchContactPhone": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "PaymentMethod": {
                  "style": "form"
                },
                "Amount": {
                  "style": "form"
                },
                "ReceiptImage": {
                  "style": "form"
                },
                "SenderPhoneNumber": {
                  "style": "form"
                },
                "DeliveryAreaId": {
                  "style": "form"
                },
                "ContactName": {
                  "style": "form"
                },
                "ContactPhone": {
                  "style": "form"
                },
                "Address": {
                  "style": "form"
                },
                "RepresentativeNotes": {
                  "style": "form"
                },
                "SlotId": {
                  "style": "form"
                },
                "DonorName": {
                  "style": "form"
                },
                "BranchContactPhone": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/subscriptions/{id}": {
      "delete": {
        "tags": [
          "Subscriptions — Donor"
        ],
        "summary": "[Donor] إلغاء اشتراك",
        "operationId": "Subscriptions_Cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSubscriptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSubscriptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/subscriptions/payments/{paymentId}/verify": {
      "post": {
        "tags": [
          "Subscriptions — Staff"
        ],
        "summary": "[Reception/Admin] تأكيد دفعة",
        "operationId": "Subscriptions_VerifyPayment",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/subscriptions/payments/{paymentId}/reject": {
      "post": {
        "tags": [
          "Subscriptions — Staff"
        ],
        "summary": "[Reception/Admin] رفض دفعة",
        "operationId": "Subscriptions_RejectPayment",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPaymentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPaymentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/subscriptions/slots": {
      "post": {
        "tags": [
          "Subscriptions — Admin"
        ],
        "summary": "[Admin] إضافة موعد متاح في المقر",
        "description": "**Request Example:**\r\n```json\r\n{\r\n  \"slotDate\": \"2026-05-01\",\r\n  \"openFrom\": \"10:00:00\",\r\n  \"openTo\": \"18:00:00\",\r\n  \"maxCapacity\": 20,\r\n  \"notes\": \"يوم الأربعاء\"\r\n}\r\n```",
        "operationId": "Subscriptions_AddSlot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlotRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlotRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Subscriptions — Admin"
        ],
        "summary": "[Admin] جلب كل المواعيد",
        "operationId": "Subscriptions_GetAllSlots",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CancelSubscriptionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateSlotRequest": {
        "type": "object",
        "properties": {
          "slotDate": {
            "type": "string",
            "format": "date-time"
          },
          "openFrom": {
            "type": "string",
            "format": "date-span"
          },
          "openTo": {
            "type": "string",
            "format": "date-span"
          },
          "maxCapacity": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateSubscriptionDto": {
        "type": "object",
        "properties": {
          "sponsorshipId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "paymentCycle": {
            "$ref": "#/components/schemas/PaymentCycle"
          }
        },
        "additionalProperties": false
      },
...
      "PaymentCycle": {
        "enum": [
          1,
          3,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "PaymentMethod": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "RejectPaymentRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "أدخل الـ Token هكذا: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}